Add the 'Wire' built-in chip.
This chip just passes through a single input pin to a single output pin
without any changes.

It's intended to be used in chip designs that need to connect an input
pin directly to an output pin without applying any gates to it.
Unfortunately because HDL needs everything to be a "part", this is the
only sensible way to make that happen.